# Update to copy unaligned data into buffer during sha256