# Description

When aarch64 optimized assembly is used, the ld1 instruction is used to directly read data from a user-supplied data buffer during the sha256 update, which may be unaligned on a 64-bit boundary especially during wolfBoot.  This correction detects non-aligned data pointers and XMEMCPY's the data into the aligned sha256->buffer structure, which roughly approximates the same operations the hardware would perform, but this avoids any unaligned 64-bit reads from memory.  This specifically corrects accesses to on-chip RAM (OCRAM) on ARMv8 processors that must be mapped as Device instead of Normal memory, as all unaligned accesses to Device memory generate a synchronous unaligned data exception.

# Testing

This change was installed on an NXP LS1028A board with wolfBoot executing out of OCRAM.  Without this correction, all boot attempts with ARMASM enabled failed as the size of the image header is 44 bytes, causing the initial image data pointer to be unaligned. 

# Checklist

 - [ None ] added tests - triggering this error would require a specific hardware setup.  No additional tests necessary.
 - [ N/A ] updated/added doxygen
 - [ N/A ] updated appropriate READMEs
 - [ N/A] Updated manual and documentation