@JacobBarthelmeh 2023-05-31
On a Cortex-A53:

Before this PR:
```
# /tmp/benchmark -sha256
------------------------------------------------------------------------------
 wolfSSL version 5.6.0
------------------------------------------------------------------------------
Math:   Multi-Precision: Wolf(SP) word-size=64 bits=4096 sp_int.c
wolfCrypt Benchmark (block bytes 1048576, min 1.0 sec each)
SHA-256 updates on pointer 359795c040
SHA-256                    575 MiB took 1.005 seconds,  572.139 MiB/s
SHA-256 updates on pointer 3597b5e041
SHA-256                    545 MiB took 1.006 seconds,  541.750 MiB/s
```

After this PR:
```
# /tmp/benchmark -sha256
------------------------------------------------------------------------------
 wolfSSL version 5.6.0
------------------------------------------------------------------------------
Math:   Multi-Precision: Wolf(SP) word-size=64 bits=4096 sp_int.c
wolfCrypt Benchmark (block bytes 1048576, min 1.0 sec each)
SHA-256 updates on pointer 439a639040
SHA-256                    575 MiB took 1.005 seconds,  572.139 MiB/s
SHA-256 updates on pointer 439a83b041
SHA-256                    185 MiB took 1.009 seconds,  183.350 MiB/s
```

Can we make this dependent on a specific architecture or macro flag?