@billphipps 2023-06-01
> On a Cortex-A53:
> 
> Before this PR:
> 
> ```
> # /tmp/benchmark -sha256
> ------------------------------------------------------------------------------
>  wolfSSL version 5.6.0
> ------------------------------------------------------------------------------
> Math:   Multi-Precision: Wolf(SP) word-size=64 bits=4096 sp_int.c
> wolfCrypt Benchmark (block bytes 1048576, min 1.0 sec each)
> SHA-256 updates on pointer 359795c040
> SHA-256                    575 MiB took 1.005 seconds,  572.139 MiB/s
> SHA-256 updates on pointer 3597b5e041
> SHA-256                    545 MiB took 1.006 seconds,  541.750 MiB/s
> ```
> 
> After this PR:
> 
> ```
> # /tmp/benchmark -sha256
> ------------------------------------------------------------------------------
>  wolfSSL version 5.6.0
> ------------------------------------------------------------------------------
> Math:   Multi-Precision: Wolf(SP) word-size=64 bits=4096 sp_int.c
> wolfCrypt Benchmark (block bytes 1048576, min 1.0 sec each)
> SHA-256 updates on pointer 439a639040
> SHA-256                    575 MiB took 1.005 seconds,  572.139 MiB/s
> SHA-256 updates on pointer 439a83b041
> SHA-256                    185 MiB took 1.009 seconds,  183.350 MiB/s
> ```
> 
> Can we make this dependent on a specific architecture or macro flag?

Yup!  I was thinking the exact same thing, but got stuck deciding between providing an ARMASM-specific flag (like WOLFSSL_ARMASM_FORCE_ALIGNMENT) versus using the global WOLFSSL_USE_ALIGN (which really only implies static alignment).   In my case, I'm leaning towards using WOLFSSL_USE_ALIGN as this would match how the XILINX offload works as well, but it could inadvertently cause performance problems in cases that don't need the alignment for these operations.  Any opinions before I push again?