@JacobBarthelmeh 2023-06-01
Looks like in types.h the macro WOLFSSL_ARMASM turns on WOLFSSL_USE_ALIGN. Maybe the macro  WC_HASH_DATA_ALIGNMENT would be a good fit here like with the MMCAU code section.